third_party/WebKit/LayoutTests/imported/csswg-test/css-snap-size-1/snap-width-parsing-001-expected.txt

This is a testharness.js-based test.
FAIL initial value assert_equals: expected (string) "0px" but got (undefined) undefined
FAIL snap-width: 20px assert_equals: expected (string) "20px" but got (undefined) undefined
FAIL snap-width should not inherit assert_equals: expected (string) "0px" but got (undefined) undefined
FAIL snap-width: inherit assert_equals: expected (string) "20px" but got (undefined) undefined
FAIL snap-width: 20px; snap-width: initial assert_equals: expected (string) "0px" but got (undefined) undefined
FAIL snap-width: 20px; snap-width: 0 assert_equals: expected (string) "0px" but got (undefined) undefined
FAIL snap-width: 20px; snap-width: 1 assert_equals: expected (string) "20px" but got (undefined) undefined
FAIL snap-width: 20px; snap-width: -1px assert_equals: expected (string) "20px" but got (undefined) undefined
FAIL snap-width: 20px; snap-width: 10% assert_equals: expected (string) "20px" but got (undefined) undefined
FAIL Computed value must be the absolute length Cannot read property 'slice' of undefined
Harness: the test ran to completion.
